requirements.txt
absl-py==0.4.0
astor==0.7.1
audioread==2.1.6
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
cycler==0.10.0
Cython==0.28.4
dask==0.18.2
decorator==4.3.0
Django==1.10
dlib==19.15.0
face-recognition==1.2.3
face-recognition-models==0.3.0
ffmpeg-python==0.1.16
future==0.16.0
gast==0.2.0
grpcio==1.12.1
h5py==2.8.0
idna==2.7
imageio==2.3.0
joblib==0.12.3
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
librosa==0.6.2
llvmlite==0.24.0
Markdown==2.6.11
matplotlib==2.2.2
mkl-fft==1.0.2
mkl-random==1.0.1
moviepy==0.2.3.5
networkx==2.1
numba==0.39.0
numpy==1.14.5
olefile==0.45.1
pandas==0.23.3
Pillow==5.1.0
protobuf==3.6.0
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
PyYAML==3.13
requests==2.19.1
resampy==0.2.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
SoundFile==0.10.2
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
toolz==0.9.0
torch==0.4.0
tornado==5.0.2
tqdm==4.24.0
urllib3==1.23
Werkzeug==0.14.1
wincertstore==0.2
